Split CacheContext
Split CacheContext unit test

Split CacheContext
